Broken anchor link in Advanced Concepts. (#1951)
The link should probably either go to directly to the beginning of main concepts (which is what my proposes change does), or to the anchor  -tag of "Step-by-Step Guide" (#step-by-step-guide). There is no other link in the article which refers to another anchor-tag, that's why I chose to directly link to the main-concepts.