Add free React course by the University of Helsinki (#3789)
Full Stack Open course is part of the computer science studies in the University of Helsinki, and also offered as a free massive open online course (MOOC). The main focus is on building single page applications with React that use REST APIs built with Node.js. It is available in English, Spanish, Chinese and Finnish.